Automatization of the creation of ping service
