+++ node/clients/index.ts
@@ -94,6 +94,8 @@ export class Clients extends IOClients {
 
 const TIMEOUT_MS = 60000
 const memoryCache = new LRUCache<string, any>({ max: 5000 })
+const TREE_SECONDS_MS = 3 * 1000
+const CONCURRENCY = 10
 
 metrics.trackCache('financial-commission', memoryCache)
 
@@ -104,6 +106,14 @@ const clients: ClientsConfig<Clients> = {
       retries: 2,
       timeout: TIMEOUT_MS,
       memoryCache,
+    },
+    events: {
+      exponentialTimeoutCoefficient: 2,
+      exponentialBackoffCoefficient: 2,
+      initialBackoffDelay: 50,
+      retries: 1,
+      timeout: TREE_SECONDS_MS,
+      concurrency: CONCURRENCY,
     },
   },
 }
